add DISABLE_MMAP flag which allows for disabling usage of mmap on Linux
This allows disabling the usage of mmap on Linux. In some deployments of Zoekt
we have seen this mmap usage lead to large variance in the response time of
Zoekt after specific indexes are not used for a few minutes and get paged out
of the OS page cache. e.g. the page cache gets invalidated, someone runs a
search query that relies on the mmap'd index, and the OS then must load a
substantial amount of data from disk resulting in a response time of e.g. 10s
instead of the regular 0.15s.

Using `DISABLE_MMAP=true` works around this problem by ensuring the index
files are always loaded into memory. Another option considered would've
been to use `madvise`, but this seemed more complex to figure out the exact
semantics of (after all, it is just advise for the kernel and not guaranteed)
especially in the context of zoekt running in e.g. Alpine linux with an arbitrary
host linux kernel.

The implementation is simple because Zoekt already supports loading index files
into memory on non-linux OSs, this just makes it optional on Linux (off by default)
where the behavior previously would've been to always use mmap.