indexdata: add new line counts to RepoStats
We want to start tracking lines of code in Sourcegraph. To do this we
extended the statistics we calculate when loading a shard in to consult
the newline index. From the newline index we are able to lookup the
number of "\n" occurences in each document.

This algorithm changes the complexity of calculateStats from O(1) to
O(n) where n is the number of documents. This is an extra n IO
operations as well (since we will seek N times through the index to
lookup the newline index for each document). Experimentally we measured
that shard load time increased by 10%. This is not a problem in most
instances, but does represent a noticeable slow down to zoekt-webserver
startup in very large instances. We believe this is worth the trade-off
to avoid introducing a lot more complexity.

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>
Change-Id: I08c1b01942f449ee1207e7d46349c85f3dc08c8f