build: enforce that repoid is up to date
We want to enforce this in our next release of Sourcegraph to ensure all
indexes have the Sourcegraph repository ID set. In the last release we
would only have the repoid set if we re-indexed. This would only occur
if the repository was updated.

This should only be landed midway through our development cycle.

Change-Id: Ibba3f436caa99c0ccdb3e1cac4d0c13bffa2ee23