shards: introduce scheduler abstraction over throttle
We want to make our throttling take into account longer running
searches. As such we introduce a lightweight scheduler and process
type. In this commit the scheduler is behaving exactly the same as the
pre-existing semaphore based throttler. We introduce this in a separate
commit to make the functional change in the next commit clearer.

Change-Id: I4e2d8a11a3c72f76fef18fa623958845d9d2fcd8