shards: introduce process.Yield
The actual yield logic is not implemented yet, but this commit adds the
necessary scaffolding to make the scheduler yielding logic self
contained in the next commit.

We document what we expect the scheduler to acheive with
yield. We only want to call yield once (when we downgrade a search), so
we introduce a deadlineTimer to help track when that should occur. We
add some benchmarks to ensure this is reasonably efficient:

  goos: darwin
  goarch: amd64
  pkg: github.com/google/zoekt/shards
  cpu: Intel(R) Core(TM) i7-8559U CPU @ 2.70GHz
  BenchmarkYield/timer-8         	349413274	        3.449 ns/op
  BenchmarkYield/now-8           	17178878	       65.88 ns/op
  BenchmarkYield/deadlineTimer-8 	252081333	        4.712 ns/op
  BenchmarkYield/yield-8         	170765276	        7.012 ns/op

Change-Id: Iabb05825fc0cd63910832a3dab3033e91a9cb773