shards: implement a batch queue
We now have two semaphores, one for interactive queries and one for
batch. The batch size is 1/4 of interactive. When we downgrade a
process (after 5s) we release the interactive semaphore and acquire the
batch semaphore. We carefully handle the case of acquiring failing
during this process.

The RWMutex is always held until our final call to Release to prevent an
exclusive process acquiring while a process is switching queues.

Change-Id: Ibd0ea99abec809a67fabcb40b410a11ad43d9557