shards: ZOEKTSCHED tuneable to disable scheduler
We introduce a scheduler interface such that we can change the scheduler
at runtime. Now if we are concerned about the new scheduler we can set
the environment variable "ZOEKTSCHED=disable=1". When set we use the old
method of throttling (a single semaphore).

We introduce ZOEKTSCHED which behaves much like the GODEBUG variable. We
do it in this manor so it is cleaner to introduce more tuneables.

Change-Id: I1301f34d479c49b6d172c6f63a257ba34e0d23cd