streaming: update buffering strategy
Up to now, we sent 1 event/shard for queries containing very common
patterns. This can potentially create a lot of network traffic and
worsen performance. With this change, we switch to a strategy that
combines shards based on a timer and a file limit.

The strategy can be summarized as follows:
- Send the first event as soon as we have found the first match.
- Send the next event either after 50ms or whenever we have found matches
in more than 100 files.