indexData: support multiple repositories (#91)
This changes `indexData` to support multiple repositories. 

Some fields of `indexData`, such as `fileBranchMask`, are document-based and
naturally support multiple repos. For other fields, such as `branchNames`, we
convert old types to slices, where each item corresponds to a repository stored
in the shard. 

As a consequence we have to update `simplify` and support those query types, which
where previously removed by simplify (Repo, RepoBranches, RepoSet).

The on-disk format of shards is unchanged, but we update `readIndexData` to read
existing shard into the new format.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>