indexdata: support List with non-constant query on compound shard
Our first stab at implementing this was incorrect. We did a search with
ShardMaxMatchCount=1. If we found a result we returned all repos on the
shard. We have to do a search per repo to correctly decide if a
repository is part of the list.

Additionally we fix a bug where we returned a reference to the loop
variable rle, which lead to all repoListEntries in the list being the
last one.

In general our uses of ShardMaxMatchCount=1 will be incorrect with
compound shards. We should consider introducing ShardRepoMaxMatchCount
option for sourcegraph to use. Alternativelly, we should use List in
cases we want a list of repositories in Sourcegraph.

Change-Id: I3eafc4ce4f68810fb399a3c17d1f45b10b96daea