indexdata: remove simple shard special casing
Simple shard special casing allows us to avoid work in the
matchtree. However, simpleMultiRepo will simplify simple shard as
well. So we can remove the special casing.

Additionally this fixes a bad bug. We passed in "in" instead of "q" to
the special casing. Not only is this wrong, it could result in us
returning a query which references itself => stack overflow.

This is an important case to cover once we start adding in better test
coverage for compound shards.

Change-Id: Ie2be0660f19e29bd9f55eee767829bd626cb9a44