indexdata: remove simple shard special casing (#100)
Simple shard special casing allows us to avoid work in the
matchtree. However, simpleMultiRepo will simplify simple shard as
well. So we can remove the special casing.

Additionally this fixes a bad bug. We passed in "in" instead of "q" to
the special casing. Not only is this wrong, it could result in us
returning a query which references itself => stack overflow.

This is an important case to cover once we start adding in better test
coverage for compound shards.