BenchmarkMinimalRepoListEncodings on a larger size
The older benchmark only had a size of 2, which isn't representative of
the issues we are seeing in production. I upped the size to 13k.

We see that encoding maps results in a lot more allocations => GC
pressure. It is also noticeably slower. However, on encoding size it is
fine. Another interesting thing to measure is decode and in-memory
representation size.

This shows we should probably add a custom encoding to avoid the badness
of gob encoding of maps.

Change-Id: I3cb08e833b6aac5188a65ddaf438f1cd6ffa32bf