indexserver: write public.txt (#88)
This commit will write a file public.txt which contains all public
repositories for this replica. The intention is that zoekt-webserver
will read this file and support filtering based on if a repository is in
this set.

We chose a simple format for public.txt to make it easy to inspect,
which we have been finding very useful in practice. We take a similiar
approach as to priorities where we handle intermittent failures in
option fetching.