indexData: hydrate with public data
The goal is to hydrate indexData with out-of-band updates of the repo
visibility. We keep public.txt in sync with shardedSearcher.public. Each
update to public.txt invalidates shardedSearcher's cache. While we fill
the new cache, we hydrate indexData based on shardedSearcher.public.

This mechanism can be used for all out-of-band hydrations of indexData.