support visibility queries
This adds atoms v:/visibility: to the query langauge.

We already save visibility data in the rawConfig of repoMetaData and we
can use that to create a docMatchTree that filters for visibility.

We add a check to incrementalSkipIndexing to make sure that repos we
haven't indexed recently will be reindexed with proper visibility.

If a repo switches from public to private, or vice versa,
incrementalSkipIndexing will return false and trigger a reindex because
the new rawConfig and the existing rawConfig in indexData don't match.
We expect this to be a rare event.