indexserver: prefer items that index in queue
Currently if an item fails to index, it normally will continue to fail
to index. We get some perverse behaviour since they will never move to
the "indexed=true" state, so they will always be at the front of the
queue.

This commit updates the priority queue to prefer items that need
indexing which are not in a failed state.

Change-Id: I533c265c27c0c0f38faec1ee5b25699998edda51