all: support mutable metadata files for shards  (#108)
We introduce a ".meta" file which is the "repoMetaData" toc section. It
marshals over the existing one if present. In v17 of our file format we
will remove repoMetadata from the toc and purely rely on the ".meta"
files.

This is how we can have mutable data alongside our shards. This main
target for this is tombstones.

There are a few other approaches to this we have debated a bit. This
approach leads to relatively minor changes to the code. The main thing I
like is the "read path" is kept simple and data related to an individual
shard remains "close" to the shard and decoupled to other shards.

Other approaches:

- single file in indexdir storing metadata for all shards. This requires
more changes. It will require extra bookkeeping if we move to a design
where shards can more easily move between replicas.

- mutating the file. This approach is appealing. We can't just mutate
due to extra complication around synchronization. So we will need some
sort of appending to end of file. This requires bigger changes to the
file format. Additionally it is useful to treat the bulk of indexdata as
immutable. However, this may be a useful approach in the future.

Change-Id: I218b69e239ce84241b62408212ea208c8490a77d