query: support rawConfig queries
This change removes visibility queries in favor of query.RawConfig queries.
query.RawConfig comes with a constructor that takes an optional mask of
flags (Public, Private, Forks, NoForks, Archived, NoArchived).

For example, to query all private forks:
```
q = query.NewRawConfig(query.Forks|query.Private)

```

The default flag is `Public|NoForks|NoArchived`.

At query time, the mask is compared to a similar mask in the
repository's meta data.

A repository's mask is created based on RawConfig when we read the
IndexFile.