indexserver: update meta file if only RawConfig changes
This commit allows us to add/update fields like public and priority
without requiring a full re-index. It uses the recently introduced
".meta" file to accomplish this.

We currently use an allowlist on fields which trigger updates. This is
what we did previously, but the list has been expanded to all RawConfig
fields Sourcegraph cares about. We didn't extend this functionality to
all fields on Repository since some fields change between builds / it
would be more work. Additionally we can't use all fields in RawConfig
since some are set by the index tools themselves.

I validated the above assumptions by patching indexserver and running it
on indexed-search-0 in production. See the patch and an example of the
Repository diff here:
https://gist.github.com/keegancsmith/0121e0cce7fb93daede78f2c0373f2cd

This commit is still missing tests. I will be adding them soon, but I
have tested manually and would appreciate a review. I may merge in and
follow-up with tests.

Change-Id: Id5e72a4a812279b7a1b4746d29a0d92d9e019e27