all: introduce Repository.MergeMutable
We use this function to decide if we can update the meta
file. Additionally it is used by IncrementalSkipIndexing to decide if
its just a meta update. This avoids us coupling logic around what
constitutes a meta update and how to do it.

We take special care to ensure some fields remain immutable, are
ignored, etc. See the code.

This was tested by running zoekt-git-index locally, mutating a
repository and observing the effects.

Change-Id: I2e1610d24c6c453374ad2f20d18a9d3ef25653c2