indexserver: update meta file if only RawConfig changes (#113)
This commit allows us to add/update fields like public and priority
without requiring a full re-index. It uses the recently introduced
".meta" file to accomplish this.

We introduce Repository.MergeMutable which will handles merging Repository
types. In particular it targets the RepositoryDescription that build.Options
creates. Using this function we can update the meta file.

I validated the above assumptions by patching indexserver and running it
on indexed-search-0 in production. See the patch and an example of the
Repository diff here:
https://gist.github.com/keegancsmith/0121e0cce7fb93daede78f2c0373f2cd

I also tested locally in Sourcegraph. I had files which started getting .meta
files with updated priorities.

This commit is still missing tests. I will be adding them soon, but I
have tested manually and would appreciate a review. I may merge in and
follow-up with tests.

Change-Id: Id5e72a4a812279b7a1b4746d29a0d92d9e019e27