build: remove old meta files when indexing (#119)
Leaving the old meta files around lead to us keeping the old meta
data. This meant we would have a new index, but we kept the old branches
information. This lead to us consistently re-indexing as well as
returning incorrect commit shas for files in results.

Test-Plan: "zoekt-git-index" on a repository with an existing .meta
file. The .meta file was deleted:

  $ ls index
  github.com%2Fkeegancsmith%2Frpc_v16.00000.zoekt      github.com%2Fkeegancsmith%2Frpc_v16.00000.zoekt.meta
  $ go run ./cmd/zoekt-git-index -index index -incremental=false ~/go/src/github.com/keegancsmith/rpc
  2021/07/21 14:06:34 finished index/github.com%2Fkeegancsmith%2Frpc_v16.00000.zoekt: 299309 index bytes (overhead 4.0)
  $ ls index
  github.com%2Fkeegancsmith%2Frpc_v16.00000.zoekt zoekt-builder-shard-log.tsv

Change-Id: I97a9581c1e092c9e460e842843ecafce3013fc2f