ctags: timeout parse requests after a minute
I believe the most common reason for a repo to not index is making ctags
hang. This commit implements a timeout which will detect and fail
indexing if ctags takes to long to respond for a single file.

We hard fail so we can notice these failures and follow-up to see what
is causing the issue. The main benefit of this change is failing an
index job much sooner (1minute), rather than waiting for indexserver to
kill it after 30 minutes.