indexserver: introduce Sourcegraph struct
All the functions moved into the Sourcegraph file are now methods of the
Sourcegraph type. This will allow us to swap it out. The intention of
this is to make a mode of indexserver which is more convenient for
manual testing and doesn't rely on the full sourcegraph dev env running.