shards: log a summary of shards in toLoad
We don't log every shard we load due to logspam on startup. However,
while zoekt-webserver is running it is useful to see more detailed
information since we are normally only loading a small number of shards
at a time.

This commit strikes a balance by only logging 5 of the shard names in
the list of shards to load.

Additionally for the unloading log line we only mention the base of the
path since it should always be in the indexdir.

Change-Id: I444a594cf5ea3359772b5f42651ee860b2a0fc1f