indexserver: expand compound shards
This is an intermediary step towards testing compound shards in
production. `getShards` will now return compound shard as
individual shards pointing to the same file/shard.

As a consequence, indexserver is now aware of the repos contained in a
compound shard and won't reindex an existing repo.

On the downside, cleanup will delete the entire compound shard if one
the repos has to be discarded. Just for the sake of testing, I propose
that we temporarily disable cleanup on one of the Zoekt instances.