shards: instrument List all RepoStat values
We are most interested in IndexBytes, which reports back how much ram
our index is taking up. However, we report everything in RepoStats since
it may be useful in the future.

The Sourcegraph frontend is regularly calling List with q=true, so this
relies on that fact. In the future if this stat stops working we can add
another way to trigger this call / calculate this stat.