eval: correct calculation for FilesSkipped in compound shards
lastDoc can be bigger than the number of documents in a specific repo
within a compound shard. As such we can end up with negative
FilesSkipped, which is incorrect. Instead we update the statistic to be
based on the number of files remaining in the whole shard.

This number has always been an upper bound. For example if you have a
filename filter, we would still return a file skipped of every remaining
file (not respecting the file filter).

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>