indexserver: compound shard support (#152)
When updating an index for a repository we need to check against any
existing index for a repository. This commit adds support for looking in
compound shards for a repository when updating. Additionally it teaches
indexserver's cleanup about compound shards.

This commit does not use tombstoning. Instead it deletes compound shards
if any repository in it needs to be removed/updated. This means we then
need to re-index everything in the compound shard. This is not our end
state, but rather an intermediate step so that we can test compound
shards without pausing the indexserver.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>