rpc: cache encoding of queries (#159)
We are paying a large cost on horizontal clients since we encode the
query per replica. This is a potential solution where we add a new query
type which will cache the marshalled value. We plan to only use this for
places we speak gob encoding. As such we unwrap GobCache in the two
places we do gob decoding: Streaming and RPC.