indexserver: port merging policy to Go
This ports the merging policy in /cmd/zoekt-merge-index from Python to
Go. I added the ported policy to zoekt-sourcegraph-indexserver because
that is were merging will finally run in background process.

To make testing easier, I added a command to indexserver that let's us
trigger a merge job manually.

After testing the merge policy manually, the next step is to run merging
every x hours on a subset of our instances in production and monitor the
behavior.