all: remove type: support (#163)
This was added a few years ago and was never used. Originally I added it
for supporting hierarchical queries in Sourcegraph. But we never shipped
the stuff I worked on on the Sourcegraph side, but the code remained in
Zoekt. This deletes likely one of the most significant divergences from
upstream.

We may want something like this again in the future. But according to
Rijnard it is hard to directly use this feature since it doesn't work
the same way and/or queries work in Sourcegraph. So best to remove.