all: switch SourcegraphID from uint32 to int32
In Sourcegraph the type RepoID is int32. This updates our code to be
consistent with Sourcegraph. This is a non-backwards compatible change
since it will change the gob encoding. As such it needs to be rolled out
with the sourcegraph frontend.

This also brings up another aspect. Should we change Sourcegraph's
RepoID type to actually be uint32. I believe that is what it really is
in the DB. That is likely a better change than this.