indexserver: make queue based on RepoID (#178)
Since we made the queue IndexOptions centered we can choose any unique
attribute as the key. The RepoID is not only smaller, but we will soon
need to add a way to "bump" a repo by id. This is where we add a repo
back to the queue with the old options.