indexserver: use repo IDs instead of names
This moves the bulk of our interactions with Sourcegraph to be based on
repository IDs instead of names. To do this required also updating our
cleanup routines to be ID based.

Cleanup is based on the list of assigned repositories. We could keep the
name based cleanup if we did it after we had resolved options. However,
resolving options can fail so I opted to update the cleanup code to be
ID based. This is where the bulk of the changes occurred. We set
tombstones as part of cleanup, so that also switched to being ID based.