indexserver: update merge policy
This updates the merge policy like follows:
- Don't merge shards by size
- Exclude large compound shards from merging

Sorting by size led to a very uneven distribution of shards to compound
shards. A more uniform distribution should lead to a more predictable
performance during search.

We introduce the threshold to exclude large compound shards from merging.
It takes around 2 minutes to create a compound shard of 2 GiB. This is
true even if we just add 1 repo to an existing compound shard. The
reason is that we don't support incremental merging, but instead create
a compound shard from scratch for each merge operation. Hence we want to
avoid merging a compound shard with other smaller candidate shards if
the compound shard already has a decent size.

The concrete value of the threshold is not important as long as it is
smaller than the targetSize and large enough to see sufficient benefits
from merging.