shards: send max pending priority while searching
On Cloud we see out of order search results on the first page. We
believe the reason might be that Zoekt doesn't report the pending
priority if it doesn't find any matches.

With this change we send the max pending priority as we search to keep
frontend informed even if we don't find matches.