don't create compound shards below target size
up to now, it was possible that we created really small compound shards
in case not enough simple shards were available to reach the target
size.

However, with this PR, vacuum deletes compound shards that are too small. This
could potentially create an endless cycle of create -> delete -> create
-> ...

Hence we update the merge strategy to make sure we only create compound
shards that are at least as big as the target size.