indexserver: Sourcegraph interface is just List (#201)
This moves a lot of the business logic out of the main loop, and instead
provides an iterate interface for the main loop to use. We intend on
changing how we batch and call out to Sourcegraph. So this change
encapsulates what we actually care about from the Sourcegraph
interface. I also think it makes the code read much more cleanly.