indexserver: ensure indexed for unchanged repositories (#203)
This fixes the potential issue introduced by fingerprinting. The issue
was that previously we would check the search configuration for all
repositories a replica is supposed to index. However, with config
fingerprinting we only checked those repositories that had changed in
Sourcegraph. Previously even if a repo hadn't changed the indexserver
would ensure that the shard would exist on disk. For example if a
previous index job failed, or the shard went missing due to an admin or
corruption.