mv testCreateEmptyShard -> createTestShard
Based on the PR comments I wanted to remove testCreateEmptyShard in
favor of reusing createEmptyShard. However, cleanup_test
requires a fine level of control over the number of shards per repo. It
feels wrong to update createEmptyShard to provide just to accomodate the
tests. Rather I think testCreateEmptyShard was a misnomer and should be
called createTestShard instead.