merging: convert feature flag to ENV variable
So far, shard merging was activated by placing a RIP file in the index
directory. However, we have rolled out the feature to the full cluster
and it makes more sense to use an ENV variable instead.

Note: Before deploying this, I have to add the new ENV variable to the
Zoekt deployment.