api: don't send IndexMetadata.LanguageMap
This mapping is purely internal from per-shard language IDs to language
names and useless for RPCs, but the IndexMetadata type is shared between
the on-disk format (as JSON) and the RPC format (with every request).

With the recent go-enry change, there are far more languages in the
mapping, and we noticed an unwanted increase in RAM usage. This is
especially wasteful as the client never even looks at the field.

Gob doesn't support tagging fields to mark them as non-exported, and a
custom encoder is weird, so instead when a shard is read from disk and
indexData is built, we set IndexMetadata.LanguageMap to nil.

This addresses sourcegraph/sourcegraph#28799

Change-Id: Ief24dfc7e40cf3abdb7b557a1d2b6f3fad5ed61f