rpc: aggregate Stats in List response (#223)
This commit adds an aggregate `Stats` field in `RepoList` which will be
used by the Sourcegraph frontend to report new line statistics without
having to list the full repo representation. We can use minimal listing,
which is already used and cached there, but now that response will
include aggregate statistics as well.