eval: re-enable result ranking
A long time ago we disabled zoekt's result ranking because in
Sourcegraph we sorted results by filename. This adds back the result
ranking which will rank results based on similarity to the query.

Additionally, in 7102177d we removed the code related to ranking based
on symbol data. This commit also partially reverts that.