eval: add ShardRepoMaxMatchCount
A common pattern we use is to search with ShardMaxMatchCount = 1 to find
all repositories with a match. However, with compound shards this breaks
since there may be multiple repositories in a shard. As such we
introduce a new option which covers that use case.

We now keep track of how many matches in a repository we have found and
skip documents once we hit that limit. This is done similarly to how we
support tombstones.

This commit is still missing a test. Additionally List can now be
updated to use this option instead of the search per repo code it
currently has.