indexserver: invadildate config fingerprint every ~15min (#229)
Just in case there is a fault in our config fingerprint logic, we fetch
all options every 10 to 20 minutes. For example when the logic first
shipped we sometimes didn't get options for new repositories.

Config fingerprint feature was motivated by the load we put on gitserver
on sourcegraph.com. We now go from doing that load from once every
minute to once every ~15min. I believe this should be acceptable. For
smaller instances this doesn't really matter since the load is tiny even
if we never used config fingerprint.