indexserver: scale configFingerprint reset time relative to corpus
We hardcoded 10 minutes. However, this lead to a noticeable impact on
Sourcegraph.com. As such we use a much more conservative reset time.

A better approach in the future would be to rather smear out the updates
over time. This would likely mean we can be less conservative in how
often we recheck options.