merging: remove obsolete maxSize
This removes the obsolete `maxSize` from the merge policy and the
command line flags.

In one of the earlier versions of the merge policy we merged compound
shards with simple shards. The parameter maxSize was used to decide
whether a compound is too big to be part of a merge operation.

However, we abandoned this idea some time ago. Effectively  `maxSize`
has no effect, because large shards are already excluded by other
criteria in `isExcluded(...)`, such as
(1) don't merge repos with multiple shards
(2) don't merge compound shards.