merging: support unsetting tombstones during cleanup (#250)
With this change tombstones are removed during cleanup if a repository is tombstoned and required to be indexed. This will speed up the recovery in case we accidentally drop a lot of indexes.

The mechanics are largely the same for shards in .trash/ and tombstoned repos. However, for the case in which a repo is tombstoned and in the .trash/ directory, the shard in the .trash/ directory takes precedence because we can assume it is newer.