wip: add current-commit oracle to gitIndex
One strategy is to have zoekt-git-index run git-diff itself. This avoids the need to create a separate API.

In this case, then all we'd need to do is ensure that we have the git data for both the old and new commits in the repository before zoekt-git-index starts processing the repo. This commit adds a first past at adding an oracle to gitIndex, which will read the shards on disk to determine the last indexed commit (and fetch that alongside the ones passed via the arguments)