Add support for JSON responses in the webserver (#240)
This commit allows using the webserver to serve JSON results so it can be used
as an API backend in a small deployment. In addition to that, it also returns
context lines to make it easy to build a simple client around the webserver.

The zoekt webserver makes it easy to launch and test a local instance.
However, it cannot be used as an API backend for custom web clients since it
only returns results in server rendered HTML pages. Additionally, introducing
context lines makes it easy to write clients that behave differently from the
default webserver.

Test Plan: Tested locally with the webserver, and passing
`?format=json&ctx=2`. All existing tests pass.