indexserver: remove deprecated GetIndexOptions
I'm trying to fix a bug where we don't retry when we fail to get index
options for a repository. The current interface is making it difficult
to fix. This removes the problematic API and makes it possible to handle
the error for a repository.

Ignoring the errors when getting index options mixed with the
fingerprint concept has lead to us forgetting about repos. The follow-up
commit to this will make us retry repos which failed.